pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "metis"
version = "0.1.0"
description = "CLI second brain that pairs with Obsidian"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.15.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "chromadb>=0.6.0",
    "pymupdf>=1.24.0",
    "httpx>=0.27.0",
    "trafilatura>=1.12.0",
]

[project.scripts]
metis = "metis.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/metis"]
